extract midiconnection out of context provider into its own component
