+++ src/components/Home.tsx
@@ -1,8 +1,9 @@
 import { h, Fragment } from 'preact';
 import { useContext } from 'preact/hooks';
 import Controller from './Controller';
-import { AppContext } from '../AppContext';
+import { AppContext } from './AppContextProvider';
 import SamplesLoader from './SamplesLoader';
+import MidiConnector from './MidiConnector';
 
 const App = () => {
   const { fetchHasError, samples, samplesAreLoading } = useContext(AppContext);
@@ -13,6 +14,7 @@ const App = () => {
       <SamplesLoader />
       {!!samples && !fetchHasError && <Controller />}
       {samplesAreLoading && <div>Loading</div>}
+      <MidiConnector />
       <footer>
         <h4>Made with ♥️ by Lø</h4>
       </footer>
