+++ src/components/SamplesLoader.tsx
@@ -3,7 +3,7 @@ import { useContext, useEffect } from 'preact/hooks';
 import { audioContext } from '../index';
 import { ACCEPTED_MIME_TYPES, BUCKET_URL } from '../constants';
 import fetchSamples from '../fetchSamples';
-import { AppContext } from '../AppContext';
+import { AppContext } from './AppContextProvider';
 
 const SamplesLoader = () => {
   const { setSamples, setSamplesAreLoading, setFetchHasError } = useContext(
